refactor TaskCreation (#88)
